conversations displaying depending on current user

+++ app/controllers/api/conversations_controller.rb
@@ -1,8 +1,11 @@
 class Api::ConversationsController < ApplicationController
+  before_action :authenticate_user, except: [:index, :show]
+
   def index
     # @conversations = current_user.conversations
     # @started_conversations = current_user.started_conversations
-    @conversations = Conversation.all
+    @conversations = Conversation.where(starter_id: current_user.id)
+    #this cannot be removed, it breaks the conversation index page
     render 'index.json.jbuilder'
   end
 
@@ -20,16 +23,21 @@ class Api::ConversationsController < ApplicationController
 
   def show
     @conversation = Conversation.find(params[:id])
-    render 'show.json.jbuilder'
+
+    if @conversation.starter_id == current_user.id
+      render 'show.json.jbuilder'
+    else
+      render json: {}, status: :unauthorized
+    end
   end
 
   def update
     @conversation = Conversation.find(params[:id])
 
-    @conversation.starter_id = params[:starter_id] || @conversation.starter_id
+    # @conversation.starter_id = params[:starter_id] || @conversation.starter_id
     @conversation.name = params[:name] || @conversation.name
 
-    if @conversation.save
+    if @conversation.starter_id == current_user.id && @conversation.save
       render 'show.json.jbuilder'
     else
       render json: { errors: @conversation.errors.full_messages }, status: :unprocessable_entity
@@ -38,9 +46,12 @@ class Api::ConversationsController < ApplicationController
 
   def destroy
     conversation = Conversation.find(params[:id])
-    conversation.destroy
 
-    render json: { message: "Successfully removed conversation" }
+    if conversation.starter_id == current_user.id
+      conversation.destroy
+      render json: { message: "Successfully removed conversation" }
+    else
+      render json: {}, status: :unauthorized
+    end
   end
-
 end
